+++ CHANGELOG.md
@@ -7,6 +7,7 @@ phpSysInfo 3.1.12
 ----------------
 
  - [ADD] BOSS, Canaima, VortexBox, KaOS and NixOS to detected distros
+ - [ADD] OpenHardwareMonitor sensor program suport
 
  - [FIX] Fixed incorrect network usage on FreeBSD